Defect/Story Number: 	Defect #82782 Description Of Defect/Story: 	TASC TEI Scoring MSS: Implement error handling and retry logic for TEI Scoring via MSS - Science item ID is getting displayed twice in TASC_Item_fact table in IRS DB Who Did Code Review: 	Somenath Chakroborti Code Files Changes: 	/dep-scoring-ctrl/src/com/ctb/lexington/db/mapper/trmapper/IrsTRItemFactMapper.java [Branch : OAS94] 	/dep-scoring-ctrl/src/com/ctb/lexington/db/mapper/tsmapper/IrsTASCItemFactMapper.java [Branch : OAS94] 	/dep-scoring-ctrl/src/com/ctb/lexington/domain/score/controller/trcontroller/StudentItemScoresController.java [Branch : OAS94] 	/dep-scoring-ctrl/src/com/ctb/lexington/domain/score/controller/tscontroller/StudentItemScoresController.java [Branch : OAS94] 	/dep-scoring-ctrl/src/maps/trfactmaps/IrsTRItemFact.xml [Branch : OAS94] 	/dep-scoring-ctrl/src/maps/tsfactmaps/IrsTASCItemFact.xml [Branch : OAS94]
Impacted Areas/Regression Areas: 	Scoring module for TASC Operational and TASC Readiness.
How The Issue Was Solved: 	The insertion logic for TASC and TASC Readiness TE Items have been changed with merge query so that existing entries get updated and no new row is inserted.
Hints To QA For Testing: 	Schedule a TASC Operational and TASC Readiness test. Complete exam in TDC. Check scoring in the IRS Database. Check the "SCORING_INVOKE_LOG" table for an erroneous roster and check the retry process for the roster is working.
Developer Comments:  	a) The OAS-DAS item mapping data has not been received yet for all the TE items. So, the scoring will be done for only the currently loaded items. For other items the item score value will be null.